Fix regex for expanding legacy html and plain formats
This regex assumed that the '.html' and '.text' extensions would
occur at the beginning of their extension groups.  Since they
are in the same group, it is not actually possible for both of
them to do this.  This prevented mailer code from loading
templates with the '.text.plain' extension, which interacted
with another buggy corner case to cause HTML emails to be sent
with 'Content-Type: text/plain'.

This commit updates the regex to properly pick up on extensions
that occur in the middle of the extension group.